manual.md: avoid external calls in Procmail recipe; bounce on error
This was brought to my attention via a question on Server Fault: https://serverfault.com/questions/1163955/how-do-i-solve-procmail-not-writing-to-home-maildir/1163956

This proposed change introduces the following changes:

* If `lookup-domain.pl` fails, we generate a bounce message, instead of silently discarding email to `/dev/null` and pretending everything is fine.
* This avoids the unnecessary external calls to test to perform string comparisons, which Procmail out of the box does very well by itself.

This only changes the manual installation instructions. I went looking for a file which would contain something similar for non-manual installations, but could not find it. Arguably if there is such a thing, it should be kept in sync.